australia-code/requirements.txt
airium==0.2.6
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arrow==1.3.0
asteval==0.9.33
attrs==24.2.0
bcp47==0.1.0
beautifulsoup4==4.12.3
bioregistry==0.11.23
blinker==1.8.2
branca==0.8.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
CFGraph==0.2.1
chardet==5.2.0
charset-normalizer==3.4.0
class_resolver==0.5.4
click==8.1.7
click-default-group==1.2.4
click-log==0.4.0
cloudpickle==3.1.0
curies==0.9.0
daff==1.3.46
dask==2024.8.0
deepdiff==6.7.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
docopt==0.6.2
duckdb==0.10.3
et_xmlfile==2.0.0
eutils==0.6.0
exceptiongroup==1.2.2
fastobo==0.12.3
flexcache==0.3
flexparser==0.3.1
folium==0.18.0
fqdn==1.5.1
fsspec==2024.10.0
funowl==0.2.3
geopandas==1.0.1
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
graphviz==0.20.3
gspread==6.1.4
gspread-formatting==1.2.0
hbreader==0.9.1
httplib2==0.22.0
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
iniconfig==2.0.0
isodate==0.7.2
isoduration==20.11.0
Jinja2==3.1.4
joblib==1.4.2
json-flattener==0.1.9
jsonasobj==1.3.1
jsonasobj2==1.0.4
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kgcl-rdflib==0.5.0
kgcl_schema==0.6.9
lark==1.2.2
linkml==1.8.5
linkml-dataops==0.1.0
linkml-map==0.3.8
linkml-renderer==0.3.1
linkml-runtime==1.8.3
locket==1.0.0
lxml==5.3.0
mapclassify==2.8.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
more-click==0.1.2
more-itertools==10.5.0
multimethod==1.10
mypy-extensions==1.0.0
narwhals==1.13.1
ndex2==3.9.0
networkx==3.2.1
num2words==0.5.13
numpy==2.0.2
oaklib==0.6.18
oauthlib==3.2.2
ols-client==0.1.4
ontodev-cogs==0.3.3
ontoportal-client==0.0.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
pandera==0.20.4
pansql==0.0.1
parse==1.20.2
partd==1.4.2
pathspec==0.12.1
pillow==10.4.0
Pint==0.24.3
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
polars==1.12.0
prefixcommons==0.1.12
prefixmaps==0.2.6
pronto==2.5.8
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJSG==0.11.10
PyJWT==2.9.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.6.1
PyShEx==0.8.1
PyShExC==0.9.1
pysolr==3.10.0
pystow==0.5.6
pytest==8.3.3
pytest-logging==2015.11.4
python-dateutil==2.9.0.post0
PyTrie==0.4.0
pytz==2024.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
quantulum3==0.9.2
ratelimit==2.2.1
rdflib==7.1.1
rdflib-jsonld==0.6.1
rdflib-shim==1.0.3
redivis==0.15.11
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
schema_automator==0.5.2
schemasheets==0.3.1
scikit-learn==1.5.2
scipy==1.13.1
semsql==0.3.3
shapely==2.0.6
ShExJSG==0.8.2
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
soupsieve==2.6
sparqlslurper==0.5.1
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.38.3
sssom==0.4.12
sssom-schema==1.0.0
streamlit==1.39.0
strsimpy==0.2.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.6
typeguard==4.4.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ucumvert==0.1.2
uri-template==1.3.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
webcolors==24.8.0
wrapt==1.16.0
xmltodict==0.13.0
xyzservices==2024.9.0
zipp==3.20.2
